LayoutTests/http/tests/misc/will-send-request-returns-null-on-redirect-expected.txt

http://127.0.0.1:8000/misc/resources/redirect-to-http-url.php - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/misc/resources/redirect-to-http-url.php, main document URL http://127.0.0.1:8000/misc/will-send-request-returns-null-on-redirect.html> redirectResponse (null)
<unknown> - didFinishLoading
http://127.0.0.1:8000/misc/resources/redirect-to-http-url.php - willSendRequest <NSURLRequest URL http://www.example.com/, main document URL http://127.0.0.1:8000/misc/will-send-request-returns-null-on-redirect.html> redirectResponse <NSURLResponse http://127.0.0.1:8000/misc/resources/redirect-to-http-url.php>
Returning null for this redirect
https://bugs.webkit.org/show_bug.cgi?id=27595
This test checks to make sure that things behave as expected when the resource load delegate returns null in response to willSendRequest.
